refactor: pull shared types into the domain layer
