+++ src/domain/audio/stem.ts
@@ -1,0 +1,7 @@
+// -- Types --------------------------------------------------------------------
+
+type Stem = "original" | "vocals" | "instrumental";
+
+// -- Exports ------------------------------------------------------------------
+
+export type { Stem };
