+++ src/domain/project/granularity.ts
@@ -1,0 +1,7 @@
+// -- Types --------------------------------------------------------------------
+
+type GranularityMode = "line" | "word";
+
+// -- Exports ------------------------------------------------------------------
+
+export type { GranularityMode };
