+++ src/domain/project/syllable-split-defaults.ts
@@ -1,0 +1,18 @@
+// -- Types --------------------------------------------------------------------
+
+interface SyllableSplitDefaults {
+  applyToAll: boolean;
+  caseInsensitive: boolean;
+}
+
+// -- Constants ----------------------------------------------------------------
+
+const DEFAULT_SYLLABLE_SPLIT_DEFAULTS: SyllableSplitDefaults = {
+  applyToAll: false,
+  caseInsensitive: false,
+};
+
+// -- Exports ------------------------------------------------------------------
+
+export { DEFAULT_SYLLABLE_SPLIT_DEFAULTS };
+export type { SyllableSplitDefaults };
